feat(http-listener): add silent option to listen method
The changes in this commit add a new `silent` option to the `listen` method of the `HTTPLister` class. This option allows the user to suppress the logging of the routes and the listening URL when starting the server.

The changes include:

- Adding a `ListenOptions` type with a `silent` property.
- Adding a `defaultOptions` static property to the `HTTPLister` class, which sets the default value of the `silent` option to `false`.
- Modifying the `listen` method to accept an optional `options` parameter of type `ListenOptions`.
- Implementing a `log` function that only logs the message if the `silent` option is `false`.
- Updating the logging statements in the `listen` method to use the `log` function.

These changes provide more flexibility to the `HTTPLister` class, allowing users to control the verbosity of the server startup process.