feat: add --cwd flag to serve command
Adds a new --cwd flag to the serve command to allow specifying the
current working directory. This is useful when the action file is not
located in the same directory as the CLI.

The changes also refactor the serve command to use the makeServerScript
function to bootstrap the server script, which simplifies the code and
makes it more reusable.