feat(metrics-client): add support for default labels
The changes add support for default labels in the `MetricsClient` class.
Previously, the `labels` property of the `MetricState` was set to an
empty object by default. Now, the `MetricsClient` constructor accepts an
optional `MetricsClientOptions` object, which can include a `labels`
property. These default labels are then merged with the labels provided
in the `bind` method, allowing users to set common labels for all
metrics.