chore: add prom-client and opentelemetry dependencies
This commit adds the `prom-client` and `@opentelemetry/api` dependencies to the project. The `prom-client` library is a Prometheus client for Node.js, which will be used for exposing metrics to a Prometheus monitoring system. The `@opentelemetry/api` is a dependency of `prom-client` and provides the OpenTelemetry API, which is a vendor-neutral API for collecting and reporting telemetry data.

Additionally, the `bintrees` and `tdigest` dependencies are added, which are required by `prom-client`.